Rename module from utils to i3extractor to better reflect purpose
